+++ tests/AggregateRootTest.php
@@ -3,11 +3,13 @@
 namespace Spatie\EventSourcing\Tests;
 
 use Illuminate\Support\Facades\Mail;
+use Spatie\EventSourcing\Exceptions\CouldNotPersistAggregate;
 use Spatie\EventSourcing\Exceptions\InvalidEloquentStoredEventModel;
 use Spatie\EventSourcing\Facades\Projectionist;
-use Spatie\EventSourcing\Snapshots\EloquentSnapshot;
 use Spatie\EventSourcing\Models\EloquentStoredEvent;
+use Spatie\EventSourcing\Snapshots\EloquentSnapshot;
 use Spatie\EventSourcing\Tests\TestClasses\AggregateRoots\AccountAggregateRoot;
+use Spatie\EventSourcing\Tests\TestClasses\AggregateRoots\AccountAggregateRootThatAllowsConcurrency;
 use Spatie\EventSourcing\Tests\TestClasses\AggregateRoots\AccountAggregateRootWithStoredEventRepositorySpecified;
 use Spatie\EventSourcing\Tests\TestClasses\AggregateRoots\Mailable\MoneyAddedMailable;
 use Spatie\EventSourcing\Tests\TestClasses\AggregateRoots\Projectors\AccountProjector;
@@ -189,7 +191,6 @@ class AggregateRootTest extends TestCase
             ->persist();
 
         $aggregateRoot->snapshot();
-
         $aggregateRoot->addMoney(100)->persist();
 
         $aggregateRootRetrieved = AccountAggregateRoot::retrieve($this->aggregateUuid);
@@ -290,5 +291,35 @@ class AggregateRootTest extends TestCase
 
             return true;
         });
+    }
+
+    /** @test */
+    public function it_will_throw_an_exception_if_the_latest_stored_version_id_is_not_what_we_expect()
+    {
+        $aggregateRoot = AccountAggregateRoot::retrieve($this->aggregateUuid);
+        $aggregateRoot->addMoney(100);
+
+        $aggregateRootInAnotherRequest = AccountAggregateRoot::retrieve($this->aggregateUuid);
+        $aggregateRootInAnotherRequest->addMoney(100);
+        $aggregateRootInAnotherRequest->persist();
+
+        $this->expectException(CouldNotPersistAggregate::class);
+        $aggregateRoot->persist();
+    }
+
+    /** @test */
+    public function it_can_allow_to_be_persisted_from_concurrent_events()
+    {
+        $aggregateRoot = AccountAggregateRootThatAllowsConcurrency::retrieve($this->aggregateUuid);
+        $aggregateRoot->addMoney(100);
+
+        $aggregateRootInAnotherRequest = AccountAggregateRootThatAllowsConcurrency::retrieve($this->aggregateUuid);
+        $aggregateRootInAnotherRequest->addMoney(100);
+        $aggregateRootInAnotherRequest->persist();
+
+        /** This line will now not throw an exception */
+        $aggregateRoot->persist();
+
+        $this->assertTestPassed();
     }
 }